Refactor toggle into custom hook.
