+++ src/components/NavBar.jsx
@@ -1,33 +1,17 @@
-import React, {useEffect, useRef, useState} from "react";
+import React from "react";
 import { Navbar, Container, Nav } from "react-bootstrap";
 import { NavLink } from "react-router-dom";
 import { useCurrentUser, useSetCurrentUser } from "../contexts/CurrentUserContext";
 import styles from "../styles/NavBar.module.css";
 import axios from "axios";
+import { useClickOutsideToggle } from "../hooks/useClickOutsideToggle";
 
 const NavBar = () => {
 
   const currentUser = useCurrentUser();
   const setCurrentUser = useSetCurrentUser();
 
-  const [expanded, setExpanded] = useState(false);
-
-  const burgerIconRef = useRef(null);
-
-  useEffect(() => {
-
-    const handleClickOutside = (event) => {
-      if (burgerIconRef.current && !burgerIconRef.current.contains(event.target)){
-        setExpanded(false);
-      }
-    };
-
-    document.addEventListener('mouseup', handleClickOutside);
-    return () => {
-      document.removeEventListener('mouseup'. handleClickOutside);
-    };
-  }, [burgerIconRef]
-  );
+  const {expanded, setExpanded, burgerIconRef} = useClickOutsideToggle();
 
 
   const handleSignOut = async () => {
